requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.13.0
django-countries==7.5
django-environ==0.9.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.4
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
StringGenerator==0.4.4
stripe==5.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
